#235: Use map object index as sign UID.
Signs don't have a UID in the mp2 file, which causes each loaded sign to
overwrite the old one in the map_objects map (because their UID was
always 0).

Changed to use the map object index as UID.
Also changed ActionToSign to use the index: signs are not tile addons,
but global map objects.