Fix right mouse split regression (#2603)
close #2392

Current logic
- Fixed right click split regression (at the risk of introducing a coding antipattern)
- When selecting a stack with 1 unit and right-clicking another stack with the same troop type, it will show the right-clicked troop's dialog info
- Ditto above but when right-clicking another stack with a different troop type. I think this one is already as it should be.

The regression issue is due to the press coming before the click so I made the press try to do click logic first. And if the click logic returns true skip the press logic. I am not too sure about this one since it feels like an antipattern to me. For left mouse click we got away with that one because we don't override `LeftMouseHold()` anywhere at the moment.