AI: support additional harmful combat spells (#2922)
Teaching AI to determine the value of debuff spells (such as Curse, Slow, Blind) and cast them.

In addition, now value will be reduced based on spell points cost. Modifier is a square root of cost normalized to a level 1 spell. For example, Chain Lighting costs 15 points, which is 5 times of a magic arrow, so value is divided by (~2.236) to avoid unnecessary mana waste.