More dialogs to exit with READY hotkey
Hero, Castle and Kingdom dialogs have only one button: `EXIT`.  Especially
when playing with one hand holding the mouse (to the right of the keyboard) it
is more convenient to exit these screens by hitting the num-pad `ENTER`.

Otherwise one has to either use the mouse to click on the button or use one's
left hand to hit `ESC`.