Removed some doubles paces in text displayed to the User, favoring single spaces, in order to be consistent with most other text.
Changed "Dwarf Bane" to "Dwarfbane" in the campaign award text to make it consistent with the in-game messages as well as to match a general pattern when using the "bane" suffix.

Made "Alliance" lowercase to be consistent with original game and most usages.

Changed "guild" to "Guild" since it is a proper noun and for consistency.

Changed "Carry-over forces" to "Carry over forces" to be consistent with original game.