ResetMixer is async by default (#4543)
ResetMixer is suffering from the same issues with music player:
for MIDI files, it will call SDL that will add a lag of 0.1 sec for every
action on MIDI files. This means that when calling ResetMixer during loading,
we waste 0.1 sec as well.

The fix is simply to have the call to Music::Reset be done in the
dedicated audio thread so it doesn't block the main thread.

We can put the "stop music" task into the stack of existing _musicTasks,
because it behaves the same way, ie the logic has to process in priority
the latest task and can discard all the previous ones.