Better heuristic for AI usage of Disrupting Ray
Relates to #4567

I changed the logic a little:
- if enemy has already defense < 4, we reduce the score of disrupting ray
(really rare corner case)
- if we have a weak army, it means we will do weak direct damage, so
it reduces the usefulness of disrupting ray. In that case I reduce the
score

If I replay the fight of #4567 it will make the AI stop spamming DR
and cast lightning bolt on the titans

NB : in theory the "enemy strength" should take into account the actual
enemy defense, that would also get lower as we cast DR on the enemy.
But the logic change I'm doing is still valid IMO