WIP: add base support for text mode
This mode is extremely useful for people with visual impairment. A special software (out of the scope of this project) should be used to process the output of the console and interpret it into a speech.

This is a prototype code and could be changed over time based of a given feedback.