Greatly speedup heavy video loading
The heaviest video is for the Original Campaign. Based on my measurements the constructor of SMKVideoSequence class was taking ~270 ms to load INTRO.SMK file. The problem was in extra calls to smk library as well as too precise memory reallocation instead of using the default power of 2 approach. After these changes the constructor takes around 120 ms.

We can do further optimizations but at least we significantly reduces video loading times.