Add mandatory requirement of h2d file being present
Since we want to bring the best experience to players and our h2d file contains some fixes over the original graphics we need to make h2d file presence to be mandatory in order to run the game. This approach will ensure that users will be using not a half-copied product.