Fix a delay in game while crossing a terrain using MIDI music
The original issue comes from the fact that SDL needs some time to process MIDI track before playing it. At the same time we were trying to get the number of allocated mixer channels but we had to wait as the same mutex were used by another thread to play the track.

The solution is to just cache the number of mixer channels and never call SDL function to get the same info. In this case we don't need to hold any mutex lock and it doesn't block environmental sound processing.

close #1979