Add initial build for Android Studio
This is a very basic setup of Android Studio build with many unsolved issues such as:
- release build requires signing
- the path in internal storage is org.libsdl.app
- source code of SDL2 and SDL2 mixer must be put under `android/app/jni` directory
- source code of fheroes2 directory subfolders must be put under `android/app/jni/src` and we shouldn't do this
- settings for the application are not set
- midi music does not work
 and etc and etc

However, you can build a debug version and run it on an Android device. I do not have one so I tested on an emulator.

All comments and recommendations are welcome as I have limited knowledge in Android development but we need to make such step to move forward.