Do not waste time to get useless monsters for an AI army
There are multiple examples when an AI hero walks with an army containing super strong stacks of monsters and at the same time Peasants. Monsters are useful in cases when they add value to overall army. If they add no value there is no need to even get them.

This change also forces heroes not to constantly accumulate army but explore the area if possible.

close #5885