Rework SDL events handling
The current implementation of SDL events is a complete mess. We do not even handle most of events and these should be disabled to avoid extra processing. Proper separation of SDL1 and SDL 2 allows us to support the code without mind-bending hacks.
The place where initialization of event types is initiated should be within core initialization code as it belongs there.