Improve rendering speed with software cursor rendering
The initial idea of sleep within the global loop is to prevent heavy CPU usage. However, with the introduction of software cursor rendering where we actually render a new frame within the global event loop, such the delay slows down the game by additional 1 ms. We shouldn't delay the event loop if we do some heavy computations within it.

relates to #6252