Do not draw multiple frames of Adventure Map
We were generating but not rendering multiple frames of Adventure Map while a hero was moving. This was a waste of resources by doing nothing from the perspective of players. This change is not ideal as I wanted to avoid making heavy changes. Also the change should potentially reduce fluctuations in FPS as these "extra" frames were rendered only when a hero moves to another tile.