Hero movement code cleanup
Small optimizations by caching values from functions and change some comments. Still the work is a mess but it is a small start. The ultimate goal is to avoid extra frame rendering where it is not needed like right now it happens when a hero completes a movement for a tile.