Fix missing logic for AI heroes to meet each other in the middle of a turn
We did not update map object cache when a hero moves so other heroes were using the old cached data and failed to find the hero if there is a need to meet each other.

On top of this the change reduces the amount of objects being processed for a possible movement. Only action objects should be considered useful for any movement.