Fix monster upgrade logic when AI was wasting resources for nothing
Previously `GetUpgradeCost()` method was returning a normal cost of a monster if no upgrade is available. This causes a problem in `Troops::UpgradeTroops()` method which was deducting resources from the kingdom but no upgrade was done. On top of this the method didn't even have a check whether the monster can be upgraded.

During AI turn when a hero is in a castle we execute `reinforceHeroInCastle()` function which internally executes the code to upgrade monsters. This of course didn't work for monsters like Wolves which don't even have any upgrades.