Expand gamepad / controller support for other platforms
close #4435

We were facing issues with controllers when software cursor mode was not set. In the past we were toggling between hardware and software cursor modes. However, now if software cursor mode is set we never set if off. In this case it is safe to enable controller support for such cases.