Reproducing original Summon Boat spell
In OG, summoning a boat by a hero is done as follows:
first, the boat belonging to the given hero is searched (each boat
belonging to a certain player, but frequently also belonging to certain hero). If such boat is not found, then a free boat belonging to this player is searched for. A free boat on the map that does not belong to anyone cannot be summoned.
Summoning a boat in fheroes2 works in a completely different way. BTW, this
may break a number of maps. It should be noted that summoning a boat in OG is also quite convenient - usually
it is most profitable to summon the boat that this particular hero used.
(not another boat, which another hero might need soon).
This PR restores the original algorithm. The only thing which I did not realized is adaptation for Save / Load (I ask for help).