Remove old Dialog::Message usage from the code
We need to standardize the code for the future changes as it is hard to do when we still use old and new code all over the place. This pull request replaces the usage of `Dialog::Message` by  `fheroes2::showStandardTextMessage`.

Additionally, there some places with `fheroes2::showStandardTextMessage` where Font::BIG was using causing unexpected appearance of buttons.