Reduce gold income and hero movement bonuses for AI
Since we added logic for marketplace usage by the AI and it has a very significant gold bonus, it is very hard to face the AI on highest difficulties. Having 60% of gold bonus is too huge. We are going to expand marketplace logic for monster purchase and upgrades so we should limit its bonuses.

This pull request also reduces movement bonus for the AI heroes on Expert difficulty to differentiate from Impossible difficulty.

These changes are subjective and you are welcome to suggest something different or even decline the proposal changes.