Shifted the name of the Necromancy skill in the secondary skill bar left by one pixel (# 7491)
The name of the Necromancy skill is so long that the text was overlapping the righthand side of the
gold border around the skill in the hero dialog. In the original game the text was centered more
to the left despite the text shadow overlapping the gold border on the left. This changes makes
the fheroes UI more accurately reproduce that of the original game.