Long Secondary skill names are now truncated
Part of PR #8058 that was dealing with centering of secondary
skill names in the hero screen. When the skill name was too
long for the block in the skill bar (which happens quite
frequently in languages other the English, which the game was
originally designed for) the names would get rendering across the
adjacent blocks and potentially overlap the text of the skill next
to them. This commit causes names that are longer than the longest
English skill name (which is "Necromancy"), to be truncated so that
they end with an elipsis ("..."). If the user wants to see the full
name they would inspect the skill by right-clicking the icon to open
the detail dialog.