Applied the effect from #7491 to the skill popup dialog
Made the skill name and skill level text centre properly. Also truncate
the text with '...' if the text is too long for the icon (happens in
some languages). The full, unmodified skill name and skill level text
can still be seen at the top of the popup dialog. Lastly limited the
text to only draw within the icon boundary so that in edge cases the
text shadow doesn't fall on the icon's gold border.