Updated tr.po
Fixed the addressed issues and reviewed the lines that I translated. Whoever did the last translation, they did lots of translation and formatting errors. Those can be ignored for now and I will be fixing them later. For now, I wanted to focus on translating the UI and some of the early game so the player can at least start their adventure.