Fix spell randomization logic for Artifacts and the Adventure Map objects
The original game does not have logic where it chooses whether a spell should be Combat or Adventure and only then does randomization based on the chosen category. Instead, it simply picks a random spell based on a level.

How it was discovered: I put 30 Shrine of the First Circle on a map and visited all of them by a hero. I counted how many spells belong to Adventure category. It was only 3 spells out of 30 instead of expected ~15.