Fix Editor buttons not being translated
close #9542

The solution is not perfect and the root cause of this case and any similar possible cases in the future is that a text translation is done implicitly for some functions while others don't. We need to avoid these confusion by removing translation code from any button generation code. Explicitly doing translation guarantees that everything works fine.