Fix object visiting marking logic
We had some "special" logic for wide objects which we were calling in an individual function. In the past we even had issues with updating their visited status which led to bugs. This pull request remove this special logic and adds a generic logic to check any object that might have multiple action object parts. This also applied for updating visited status for allies.

Before the changes we had no issues with these objects. However, if in the future (and most likely it's going to happen) we are going to add more objects with several action object parts then the current code is not flexible at all.