Save days spent value for each scenario of campaign
We are still missing a feature to review completed scenarios in a campaign. In the original game it is possible to do and for each scenario it displays the number of days spent in relation to the scenario. This change saves the same information to be used in the future UI updates.

relates to #4381