Limit resolutions only in software rendering mode
With the current implementation if software rendering is off and a user goes to a full-screen mode, selects lower resolution (for example, 640x480) there is no way to select a higher resolution after. This is because the operating system resolution is set to it and since we limit all resolution to the system we have no way to change it.

This pull request limits the original logic only for software rendering mode.