Store map information in a compressed format for history manager
Based on the initial tests the engine can eat up 1 GB of RAM with the old implementation of storing 2 copies of map format. With this approach the RAM usage is reduced in 10 times.

Therefore, we are also increasing the number of remembered actions for map makers.

relates to #10398

Performance wise the solution is slightly slower than the original but gives enormous memory usage reduction.